### 2024-04-03 #6577 @editorialbot
[PRE REVIEW]: GO-RXR: Global Optimization of Resonant X-ray Reflectometry
**Submitting author:** @jpcurbelo (<a href="http://orcid.org/0000-0003-2417-071X">Jesus P. Curbelo</a>)
**Repository:** https://github.com/lucaskorol21/GO-RXR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
