### 2024-04-05 8ba2c67
Added contributing, doi, license, and updated readme file
