+++ CONTRIBUTING.md
@@ -9,7 +9,7 @@ We welcome contributions from all members of the community. Here are a few ways 
 1.  **Bug Reports**: If you've come across some behavior that you think is a bug or a missing feature, the first step is to check if it's already known. For this, take a look at the [repository issues page](https://github.com/lucaskorol21/GO-RXR/issues). If you can't find anything related there, you're welcome to open a new issue.
 2.  **Feature Requests**: If you have ideas for new features or enhancements, feel free to [open a new issue](https://github.com/lucaskorol21/GO-RXR/issues) to discuss them.
 3.  **Code Contributions:** If you're interested in contributing code, please fork the repository, make your changes, and submit a pull request. Make sure to follow our coding style as possible.
-4. ****Documentation**: Improving documentation is always appreciated! If you notice any gaps or errors in the code documentation, tutorials, or user guide, please submit a pull request with your proposed changes.
+4. **Documentation**: Improving documentation is always appreciated! If you notice any gaps or errors in the code documentation, tutorials, or user guide, please submit a pull request with your proposed changes.
 
 And if you find the project valuable but don't have the time to contribute directly, there are still ways to support us and show your appreciation:
 
