### 2024-04-05 #6577 @jpcurbelo
> @jpcurbelo Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
>     * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
>
>     * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).
>
>     * [ ]  The license for this project is not clearly recognised for your submission. In addition the license file detected is not an OSI approved license. Please add a plain text `LICENSE` file to your main folder which should be an unedited OSI approved license.
>
>     * [ ]  Please format the paper as per our guidelines and have a separate Summary and Statement of Need section.

Dear @Kevin-Mattheus-Moerman, we have addressed the mentioned points. The two references lacking DOIs are a MSc Thesis and a GitHub code repository.

Here's a list of potential reviewers for our paper:
- kstenio
- marcocamma
- larsenkg
- ajinkyakadu
- CFGrote
- stuartcampbell

Please let us know if there's anything else needed.
