### 2024-07-03 43fdc09
Added solution to install reflectivity - working with python 3.10
