+++ UTILS/Pythonreflectivity.cpp
@@ -4,19 +4,19 @@
 {
     "distutils": {
         "depends": [
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include\\numpy\\arrayobject.h",
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include\\numpy\\arrayscalars.h",
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include\\numpy\\ndarrayobject.h",
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include\\numpy\\ndarraytypes.h",
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include\\numpy\\ufuncobject.h"
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include/numpy/arrayobject.h",
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include/numpy/arrayscalars.h",
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include/numpy/ndarrayobject.h",
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include/numpy/ndarraytypes.h",
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include/numpy/ufuncobject.h"
         ],
         "include_dirs": [
-            "C:\\Users\\lsk601\\PycharmProjects\\MaterialReflection\\venv\\lib\\site-packages\\numpy\\core\\include"
+            "/mnt/data/USASK_research/Reflectometry_GO-RXR/GO-RXR/venv-gorxr310/lib/python3.10/site-packages/numpy-1.26.3-py3.10-linux-x86_64.egg/numpy/core/include"
         ],
         "language": "c++",
         "name": "Pythonreflectivity",
         "sources": [
-            "Pythonreflectivity.pyx"
+            "UTILS/Pythonreflectivity.pyx"
         ]
     },
     "module_name": "Pythonreflectivity"
@@ -1520,10 +1520,10 @@ static const char *__pyx_filename;
 /* #### Code section: filename_table ### */
 
 static const char *__pyx_f[] = {
-  "Pythonreflectivity.pyx",
+  "UTILS/Pythonreflectivity.pyx",
   "<stringsource>",
[…]
