+++ setup.py
@@ -3,8 +3,8 @@ from setuptools import setup, find_packages
 setup(
     name='GO-RXR',
     version='1.0',
-    # packages=find_packages(include=['Magnetic_Scattering_Factor', 'Scattering_Factor', 'Testing', 'Tutorial']),
-    packages=['Magnetic_Scattering_Factor', 'Scattering_Factor', 'Testing', 'Tutorial'],
+    packages=find_packages(include=['Magnetic_Scattering_Factor', 'Scattering_Factor', 'Testing', 'Tutorial']),
+    # packages=['Magnetic_Scattering_Factor', 'Scattering_Factor', 'Testing', 'Tutorial'],
     url='https://github.com/lucaskorol21/GO-RXR',
     author='Lucas Korol',
     author_email='lsk601@usask.ca',
