+++ UNKNOWN.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+LICENSE
+README.md
+setup.cfg
+setup.py
+setup_reflectivity.py
+UNKNOWN.egg-info/PKG-INFO
+UNKNOWN.egg-info/SOURCES.txt
+UNKNOWN.egg-info/dependency_links.txt
+UNKNOWN.egg-info/top_level.txt
+UTILS/Pythonreflectivity.cpp
