### 2024-07-31 #6577 @JohannesBuchner
Yes, I can review this.

While checking out what this is, I already came across a few items I noticed.

The install is a little unorthodox, why not put the cython install into setup.py, that way you can also deploy to pypi and pull in all the dependencies, making it easier for people to install. For an example, see setup.py in https://github.com/JohannesBuchner/UltraNest/
Is this because many people would edit the package code and are not used to `python setup.py build_ext`?

If you store screenshots in a DOCS/ subfolder, you can embed them in the README without relying on an external provider.

It also seems to me that the non-GUI unit tests could be run through github actions, which would increase confidence in the development process.

The checklist will ask me "Do the authors describe how this software compares to other commonly-used packages?" and I may need help with finding this in the current draft.

If I read correctly, a main contribution here is the GUI and the global optimization algorithms. Maybe then it is worth stating which optimization algorithms these are (or a list of the most important ones with refs), and show the GUI.
