### 2024-08-31 3d92f3e
Create go-rxr-tests.yml
This workflow automates the testing process for the GO-RXR project, ensuring that all non-GUI unit tests are consistently executed on each push or pull request to main and version-joss-rev branches. This integration helps maintain code quality and detect issues early in the development process.

Signed-off-by: Jesus Perez Curbelo <78277407+jpcurbelo@users.noreply.github.com>
