### 2024-08-31 189ee07
Update go-rxr-tests.yml
Signed-off-by: Jesus Perez Curbelo <78277407+jpcurbelo@users.noreply.github.com>
