+++ .github/workflows/go-rxr-tests.yml
@@ -42,5 +42,5 @@ jobs:
 
     - name: Run tests
       run: |
-        python -m unittest discover -s TESTS -p "*test_data_fitting.py"
-        python -m unittest discover -s TESTS -p "*test_data_structure.py"
+        python -m unittest discover -s TESTS -p "test_data_fitting.py"
+        python -m unittest discover -s TESTS -p "test_data_structure.py"
