### 2024-08-31 297fe6b
testing print(os.listdir(os.getcwd()))
