### 2024-08-31 8605fa2
my_path = find_root_dir('GO-RXR') + '/TESTS/test_data/' + filename
