+++ TESTS/test_data_fitting.py
@@ -3,9 +3,7 @@ import copy
 import unittest
 import sys 
 
-from utils import get_test_data_path
-# Import ROOT_DIR from the __init__.py file
-from UTILS import ROOT_DIR
+from utils import find_root_dir
 
 # Get the parent directory of the current script's directory
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
@@ -33,18 +31,24 @@ class TestDataFitting(unittest.TestCase):
             my_path = os.getcwd() + '/test_data/' + filename
             script_name = '/test_data/test_script.txt'
         else:
-            my_path = os.getcwd() + '/test_data/' + filename
+            # my_path = os.getcwd() + '/test_data/' + filename
+            # script_name = '/test_data/test_script.txt'
+            my_path = find_root_dir('GO-RXR') + '/TESTS/test_data/' + filename
             script_name = '/test_data/test_script.txt'
-            # my_path = get_test_data_path(filename)
-            # script_name = get_test_data_path('test_script.txt')
 
-        print('test_data_path', os.getcwd())
-        print(os.listdir(os.getcwd()))
+        # print('test_data_path', os.getcwd())
+        # print(os.listdir(os.getcwd()))
 
-        print("ROOT_DIR", ROOT_DIR)
-        print('\nmy_path', my_path)
+        root_dir = find_root_dir('GO-RXR')
+        print('root_dir', root_dir)
+
+        print('\n', os.getcwd() + '/test_data/' + filename)
+        print(os.getcwd() + '/test_data/test_script.txt')
+
+        print('my_path', my_path)
         print('script_name', script_name)
-        aux = input('Press enter to continue')
+
+        # aux = input('Press enter to continue')
 
         sample = ds.ReadSampleHDF5(my_path)
 
