+++ TESTS/utils.py
@@ -2,17 +2,28 @@
 # This file contains utility functions for the tests.
 import os
 
-def get_test_data_path(filename):
+def find_root_dir(target_dir_name):
     """
-    Returns the full path to a test data file, correcting for different environments.
+    Recursively finds the root directory of the project by looking for a directory
+    with the specified name ('target_dir_name') starting from the current working directory.
+
+    Args:
+    - target_dir_name (str): The name of the directory to find.
+
+    Returns:
+    - str: The absolute path to the root directory if found, otherwise None.
     """
-    base_dir = os.path.dirname(os.path.abspath(__file__))
-    # Check if the path includes a nested directory like 'GO-RXR/GO-RXR'
-    if "GO-RXR/GO-RXR" in base_dir:
-        base_dir = base_dir.replace("/GO-RXR/GO-RXR", "/GO-RXR")
-    
-    test_data_dir = os.path.join(base_dir, 'test_data')
-    return os.path.join(test_data_dir, filename)
+    current_dir = os.getcwd()
+
+    while True:
+        if os.path.basename(current_dir) == target_dir_name:
+            return current_dir
+        parent_dir = os.path.dirname(current_dir)
+        # Check if we have reached the root of the filesystem
+        if parent_dir == current_dir:
+            print(f"Root directory '{target_dir_name}' not found")
+            return None
+        current_dir = parent_dir
 
 
 
