+++ TESTS/test_data_fitting.py
@@ -110,675 +110,676 @@ class TestDataFitting(unittest.TestCase):
         # test orbital
         self.assertListEqual(orbitals['Mn2'], orbitals_new['Mn2'])
 
-    def test_ChangeSampleParams_compound(self):
-        # Tests to element fit
-        filename = 'Pim4uc_test.h5'
-        if os.getcwd().split('\\')[-1] == 'Testing':
-            my_path = os.getcwd() + '/test_data/' + filename
-            script_name = '/test_data/test_script.txt'
-        else:
-            my_path = os.getcwd() + '/test_data/' + filename
-            script_name = '/test_data/test_script.txt'
-
-        sample = ds.ReadSampleHDF5(my_path)
-
-        data, data_dict, sim_dict = ds.ReadDataHDF5(my_path)
-
-        parameters = [[0, 'STRUCTURAL', 'COMPOUND','ROUGHNESS',0],[1, 'STRUCTURAL', 'COMPOUND','THICKNESS',1],
-                      [2, 'STRUCTURAL', 'COMPOUND','DENSITY',2], [3, 'STRUCTURAL', 'COMPOUND','ROUGHNESS',0],
-                      [4, 'STRUCTURAL', 'COMPOUND','THICKNESS',1], [5, 'STRUCTURAL', 'COMPOUND','DENSITY',1],
-                      [6, 'STRUCTURAL', 'COMPOUND','ROUGHNESS',0], [7, 'STRUCTURAL', 'COMPOUND','THICKNESS',1]]
-
-
-
-        x = [1.5, 1.95, 0.074, 4, 3.5, 0.025, 2.15, 12]  # parameter values
-
-        # create the background shift and scaling factor keys
-        backS = dict()
-        scaleF = dict()
-        for name in list(data_dict.keys()):
-            backS[name] = 0
-            scaleF[name] = 1
-
-        my_script, problem, my_error = checkscript(sample, fname=script_name)  # load in the script
-        orbitals = {'Mn2': [0, 0, 0, 0]}  # orbitals
-
-        sample_new, backS_new, scaleF_new, orbitals_new = changeSampleParams(x, parameters, sample, backS, scaleF,
-                                                                             my_script, orbitals)
-
-        # test sample parameters
-        sample_solution = copy.deepcopy(sample)
-        sample_solution.structure[0]['A'].roughness = 1.5
-        sample_solution.structure[0]['Ti'].roughness = 3.295897580593646
-        sample_solution.structure[0]['O'].roughness = 3.295897580593646
-        sample_solution.structure[1]['A'].thickness = 1.95
-        sample_solution.structure[1]['Ti'].thickness = 1.95
-        sample_solution.structure[1]['O'].thickness = 1.95
-        sample_solution.structure[2]['A'].density = 0.01799999999999999
-        sample_solution.structure[2]['Ti'].density = 0.01799999999999999
-        sample_solution.structure[2]['O'].density = 0.074
-        sample_solution.structure[3]['A'].roughness = 4
-        sample_solution.structure[3]['Mn'].roughness = 4.0
-        sample_solution.structure[3]['O'].roughness = 4.75
-        sample_solution.structure[4]['A'].thickness = 3.5
-        sample_solution.structure[4]['Mn'].thickness = 3.5
-        sample_solution.structure[4]['O'].thickness = 3.5
-        sample_solution.structure[5]['A'].density = 0.025
-        sample_solution.structure[5]['Mn'].density = 0.025
-        sample_solution.structure[5]['O'].density = 0.07500000000000001
-        sample_solution.structure[6]['A'].roughness = 2.15
[…]
