### 2024-08-31 3c350c8
added testing=True and script_path
