### 2024-08-31 e8d55cf
added testing=True and script_path 2
