### 2024-09-01 5ce89b1
Tests fitting and structure ok?
