+++ TESTS/test_data_fitting.py
@@ -3,8 +3,6 @@ import copy
 import unittest
 import sys 
 
-from utils import find_root_dir
-
 # Get the parent directory of the current script's directory
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 # Add the parent directory to the system path
@@ -14,6 +12,7 @@ from UTILS import (
     material_structure as ms,
     data_structure as ds,
     global_optimization as go,
+    ROOT_DIR,
 )
 from UTILS.global_optimization import changeSampleParams
 from GUI_GO import checkscript
@@ -23,37 +22,21 @@ from GUI_GO import checkscript
 # into the terminal
 
 class TestDataFitting(unittest.TestCase):
-    
+
+    def __init__(self, methodName: str = "runTest") -> None:
+        super().__init__(methodName)
+
+        self.root_dir = ROOT_DIR
+        self.filename = 'Pim4uc_test.h5'
+        self.my_path = self.root_dir + '/TESTS/test_data/' + self.filename
+        self.script_path = self.root_dir + '/TESTS/test_data/test_script.txt'
+
     def test_ChangeSampleParams_element(self):
+
         # Tests to element fit
-        filename = 'Pim4uc_test.h5'
-        if os.getcwd().split('\\')[-1] == 'Testing':
-            my_path = os.getcwd() + '/test_data/' + filename
-            script_name = '/test_data/test_script.txt'
-        else:
-            # my_path = os.getcwd() + '/test_data/' + filename
-            # script_name = '/test_data/test_script.txt'
-            root_dir = find_root_dir('GO-RXR')
-            my_path = root_dir + '/TESTS/test_data/' + filename
-            script_path = root_dir + '/TESTS/test_data/test_script.txt'
+        sample = ds.ReadSampleHDF5(self.my_path)
 
-        # print('test_data_path', os.getcwd())
-        # print(os.listdir(os.getcwd()))
-
-        root_dir = find_root_dir('GO-RXR')
-        print('root_dir', root_dir)
-
-        print('\n', os.getcwd() + '/test_data/' + filename)
-        print(os.getcwd() + '/test_data/test_script.txt')
-
-        print('my_path', my_path)
-        print('script_name', script_path)
-
-        # aux = input('Press enter to continue')
-
-        sample = ds.ReadSampleHDF5(my_path)
-
-        data, data_dict, sim_dict = ds.ReadDataHDF5(my_path)
+        data, data_dict, sim_dict = ds.ReadDataHDF5(self.my_path)
 
         parameters = [[1, 'STRUCTURAL', 'ELEMENT', 'A', 'THICKNESS'], [3, 'STRUCTURAL', 'ELEMENT', 'A', 'THICKNESS'],
                       [2, 'STRUCTURAL', 'ELEMENT', 'Ti', 'DENSITY'], [4, 'STRUCTURAL', 'ELEMENT', 'O', 'DENSITY'],
@@ -70,7 +53,7 @@ class TestDataFitting(unittest.TestCase):
             scaleF[name] = 1
 
         # my_script, problem, my_error = checkscript(sample, fname=script_name)  # load in the script
-        my_script, problem, my_error = checkscript(sample, fname=script_path, testing=True)  # load in the script
+        my_script, problem, my_error = checkscript(sample, fname=self.script_path, testing=True)  # load in the script
         orbitals = {'Mn2': [0, 0, 0, 0]}  # orbitals
 
         sample_new, backS_new, scaleF_new, orbitals_new = changeSampleParams(x, parameters, sample, backS, scaleF, my_script, orbitals)
[…]
