### 2024-09-01 9e861c5
Updated all tests to be run on GitHub actions
