### 2024-09-01 8c2e83d
Update go-rxr-tests.yml
Signed-off-by: Jesus Perez Curbelo <78277407+jpcurbelo@users.noreply.github.com>
