+++ .github/workflows/go-rxr-tests.yml
@@ -44,3 +44,7 @@ jobs:
       run: |
         python -m unittest discover -s TESTS -p "test_data_fitting.py"
         python -m unittest discover -s TESTS -p "test_data_structure.py"
+        python -m unittest discover -s TESTS -p "test_density.py"
+        python -m unittest discover -s TESTS -p "test_material_model.py"
+        python -m unittest discover -s TESTS -p "test_material_structure.py"
+        python -m unittest discover -s TESTS -p "test_reflectivity.py"
