### 2024-09-01 2a8f802
Added paragraph on optimization to paper
