+++ PAPER/paper.bib
@@ -62,3 +62,14 @@
     howpublished = {\url{https://github.com/malaclypseII/PyXMRTool/tree/master}},
     commit = {30d1ed552cf807d3635d2e274dd89151c2a83090}
 }
+
+@article{virtanen2020_scipy,
+    title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+    author={Virtanen, P. and Gommers, R. and Oliphant, T.E. and others},
+    journal={Nature Methods},
+    volume={17},
+    pages={261--272},
+    year={2020},
+    doi={10.1038/s41592-019-0686-2},
+    url={https://doi.org/10.1038/s41592-019-0686-2}
+}
