### 2024-09-02 9292cd6
Paper: added comments and figs for rev1 comments
