+++ PAPER/paper.bib
@@ -63,6 +63,42 @@
     commit = {30d1ed552cf807d3635d2e274dd89151c2a83090}
 }
 
+@article{storn1997_de,
+    title={Differential Evolution – A Simple and Efficient Heuristic for Global Optimization over Continuous Spaces},
+    author={Storn, R. and Price, K.},
+    journal={Journal of Global Optimization},
+    volume={11},
+    number={4},
+    pages={341--359},
+    year={1997},
+    doi={10.1023/A:1008202821328},
+    url={https://doi.org/10.1023/A:1008202821328}
+}
+
+@article{endres2018_shgo,
+    title={A simplicial homology algorithm for Lipschitz optimisation},
+    author={Endres, S.C. and Sandrock, C. and Focke, W.W.},
+    journal={Journal of Global Optimization},
+    volume={72},
+    number={2},
+    pages={181--217},
+    year={2018},
+    doi={10.1007/s10898-018-0627-y},
+    url={https://doi.org/10.1007/s10898-018-0627-y}
+}
+
+@article{xiang1997_dual_annealing,
+    title={Generalized Simulated Annealing Algorithm and Its Application to the Thomson Model},
+    author={Xiang, Y. and Sun, D.Y. and Fan, W. and Gong, X.G.},
+    journal={Physics Letters A},
+    volume={233},
+    number={3},
+    pages={216--220},
+    year={1997},
+    doi={10.1016/S0375-9601(97)00474-X},
+    url={https://doi.org/10.1016/S0375-9601(97)00474-X}
+}
+
 @article{virtanen2020_scipy,
     title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
     author={Virtanen, P. and Gommers, R. and Oliphant, T.E. and others},
