### 2024-09-02 a60a343
Added comparison to paper
