+++ PAPER/paper.bib
@@ -109,3 +109,50 @@
     doi={10.1038/s41592-019-0686-2},
     url={https://doi.org/10.1038/s41592-019-0686-2}
 }
+
+@article{bjorck2007_genx,
+    title={GenX: an extensible X-ray reflectivity refinement program utilizing differential evolution},
+    author={Björck, M. and Andersson, G.},
+    journal={Journal of Applied Crystallography},
+    volume={40},
+    pages={1174--1178},
+    year={2007},
+    doi={10.1107/S0021889807045086},
+    url={https://doi.org/10.1107/S0021889807045086}
+}
+
+@article{nelson2006_motofit,
+    title={Co-refinement of multiple-contrast neutron/X-ray reflectivity data using MOTOFIT},
+    author={Nelson, A.},
+    journal={Journal of Applied Crystallography},
+    volume={39},
+    pages={273--276},
+    year={2006},
+    doi={10.1107/S0021889806005073},
+    url={https://doi.org/10.1107/S0021889806005073}
+}
+
+@article{nelson2014_refl1d,
+    title={Refl1D: Software for Modeling and Fitting Reflectometry Data},
+    author={Nelson, A. R. J., et al.},
+    journal={SoftwareX},
+    volume={1-2},
+    pages={2--7},
+    year={2014},
+    doi={10.1016/j.softx.2014.06.002},
+    url={https://doi.org/10.1016/j.softx.2014.06.002}
+}
+
+@article{durniak2015_bornagain,
+    title={BornAgain: Software for Simulating and Fitting Reflectometry and Grazing-Incidence Small-Angle Scattering},
+    author={Durniak, C. and others},
+    journal={Computational Materials Science},
+    volume={103},
+    pages={78--90},
+    year={2015},
+    doi={10.1016/j.commatsci.2015.02.041},
+    url={https://doi.org/10.1016/j.commatsci.2015.02.041}
+}
+
+
+
