### 2024-09-02 abc2f9a
Corrected references in comparison section
