+++ PAPER/paper.bib
@@ -132,26 +132,24 @@
     url={https://doi.org/10.1107/S0021889806005073}
 }
 
-@article{nelson2014_refl1d,
-    title={Refl1D: Software for Modeling and Fitting Reflectometry Data},
-    author={Nelson, A. R. J., et al.},
-    journal={SoftwareX},
-    volume={1-2},
-    pages={2--7},
-    year={2014},
-    doi={10.1016/j.softx.2014.06.002},
-    url={https://doi.org/10.1016/j.softx.2014.06.002}
+@misc{refl1d_github,
+    title={{Refl1D: Reflectometry Data Analysis}},
+    author={Kienzle, P.A. and Maranville, B.B. and O'Donovan, K.V. and Ankner, J.F. and Berk, N.F. and Majkrzak, C.F.},
+    howpublished={\url{https://github.com/reflectometry/refl1d}},
+    year={2017},
+    note={GitHub repository, Accessed: 2024-09-01},
 }
 
-@article{durniak2015_bornagain,
-    title={BornAgain: Software for Simulating and Fitting Reflectometry and Grazing-Incidence Small-Angle Scattering},
-    author={Durniak, C. and others},
-    journal={Computational Materials Science},
-    volume={103},
-    pages={78--90},
-    year={2015},
-    doi={10.1016/j.commatsci.2015.02.041},
-    url={https://doi.org/10.1016/j.commatsci.2015.02.041}
+@article{pospelov2020_bornagain,
+    title={{\it BornAgain}: software for simulating and fitting grazing-incidence small-angle scattering},
+    author={Pospelov, Gennady and Van Herck, Walter and Burle, Jan and Carmona Loaiza, Juan M. and Durniak, C{\'{e}}line and Fisher, Jonathan M. and Ganeva, Marina and Yurov, Dmitry and Wuttke, Joachim},
+    journal={Journal of Applied Crystallography},
+    volume={53},
+    number={1},
+    pages={262--276},
+    year={2020},
+    doi={10.1107/S1600576719016789},
+    url={https://doi.org/10.1107/S1600576719016789},
 }
 
 
