### 2024-09-02 #6577 @jpcurbelo
> Yes, I can review this.
>
> While checking out what this is, I already came across a few items I noticed.
>
> The install is a little unorthodox, why not put the cython install into setup.py, that way you can also deploy to pypi and pull in all the dependencies, making it easier for people to install. For an example, see setup.py in https://github.com/JohannesBuchner/UltraNest/ Is this because many people would edit the package code and are not used to `python setup.py build_ext`?
>
> If you store screenshots in a DOCS/ subfolder, you can embed them in the README without relying on an external provider.
>
> It also seems to me that the non-GUI unit tests could be run through github actions, which would increase confidence in the development process.
>
> The checklist will ask me "Do the authors describe how this software compares to other commonly-used packages?" and I may need help with finding this in the current draft.
>
> If I read correctly, a main contribution here is the GUI and the global optimization algorithms. Maybe then it is worth stating which optimization algorithms these are (or a list of the most important ones with refs), and show the GUI.

Hi @JohannesBuchner,

Thank you for your detailed feedback. Here’s an update on the points you raised:

1-Installation Process: We are still working on integrating the Cython installation into setup.py. This is new territory for us, so we're taking our time to fully understand the process before making any changes. We appreciate your patience on this.

2-Screenshots in Documentation: The DOCS folder currently houses all tutorials and documentation, which include numerous screenshots. However, we've added two key screenshots (Fig 1 and Fig 3) directly into the README as part of your suggestion.

3-GitHub Actions for Unit Tests: This is now implemented, and the non-GUI unit tests are running through GitHub Actions.

4-Comparison with Other Software: We've added a paragraph at the beginning of the Comparison section, including references and comments on related work.

5-GUI and Optimization Algorithms: We've added a paragraph with references and included a figure at the end of the Basic Concepts subsection to highlight the key optimization algorithms and showcase the GUI.

Thanks again, and please let us know if there's anything else you'd like us to address.
