### 2024-09-02 d385288
Put the cython install into setup.py
