### 2024-09-02 839f966
Integrated cython install to setup and updated readme
