+++ GO_RXR.egg-info/requires.txt
@@ -1,8 +1,0 @@
-Cython==3.0.8
-numpy==1.26.3
-h5py==3.10.0
-numba==0.59.0
-pyqtgraph==0.13.3
-scipy==1.12.0
-matplotlib==3.8.1
-PyQt5==5.15.2
