+++ README.md
@@ -59,22 +59,26 @@ $ git clone https://github.com/lucaskorol21/GO-RXR.git
 We recommend creating a virtual enviromnment:
 
 ```bash
-$ virtualvenv venv-go-rxr
+$ virtualenv -p python3.10 venv-go-rxr
 ```
 
-Install the python libraries by running the setup file:
+Install the Python libraries using `pip`
 
 ```bash
-$ python setup.py install
+(venv-go-rxr) $ pip install .
 ```
 
-* First run ```$ pip install --upgrade pip setuptools``` if needed.
+This command will handle the installation of all dependencies, including compiling any Cython extensions.
 
 If the setup file does not work, then the libraries in the `requirements.txt` file can be installed.
 
+```bash
+(venv-go-rxr) $ pip install -r requirements.txt
+```
+
 #### For `matplotlib`, ensure that you have `Pilow` installed
 ```bash
-$ pip install Pillow
+(venv-go-rxr) $ pip install Pillow
 ```
 
 #### Resolving `PyQt5` Conflicts.
@@ -83,63 +87,19 @@ If you encounter conflicts with the PyQt5 package during installation or usage, 
 
 ```bash
 # Install necessary dependencies
-sudo apt-get install pyqt5-dev libqt5multimedia5-plugins
+$ sudo apt-get install pyqt5-dev libqt5multimedia5-plugins -y
 
 # Remove existing PyQt5 installations from the virtual environment
-sudo rm -f -r /usr/lib/python3/dist-packages/PyQt5 /path/to/your/virtualenv/lib/python3.x/site-packages/PyQt5
+$ sudo rm -f -r /usr/lib/python3/dist-packages/PyQt5 /path/to/your/virtualenv/lib/python3.x/site-packages/PyQt5
 
 # Create a symbolic link from the OS libraries to the virtual environment
-sudo ln -f -s /usr/lib/python3/dist-packages/PyQt5 /path/to/your/virtualenv/lib/python3.x/site-packages/PyQt5
+$ sudo ln -f -s /usr/lib/python3/dist-packages/PyQt5 /path/to/your/virtualenv/lib/python3.x/site-packages/PyQt5
 ```
 Replace /path/to/your/virtualenv with the path to your virtual environment directory and 3.x with the appropriate Python version (e.g., 3.10, 3.9, etc.). These commands aim to ensure that the global version of PyQt5 matches the one specified in your setup file by using the operating system's libraries and creating a symbolic link accordingly.
 
-#### 3. Install Python reflectivity by running
-```bash
-$ python setup_reflectivity.py install
-```
-
-In case there you found an error related to `'x86_64-linux-gnu-gcc'` permission, use
-```bash
-$ sudo python setup_reflectivity.py install
-```
-or
-```bash
-# Ensure that you have write permissions for the dist directory by running
-$ ls -ld dist/
-
-# If the ownership of the dist directory is incorrect, you can change it using the following command:
-$ sudo chown -R $USER dist/
-
-# Try running the installation command again
-$ sudo python setup_reflectivity.py install
-```
-
-If the issue persists, try the following:
-```bash
-# Remove the build directory
-$ rm -rf build/
-
-# Make sure your user has write permissions for the entire project directory
-$ sudo chown -R $USER .
-
-# Try running the installation command again
-$ python setup_reflectivity.py install
-```
-
-or 
-
-```bash
-# Install Python Development Headers
-$ sudo apt-get update
-$ sudo apt-get install python3.10-dev
-
-# Install your package using setup.py
-$ python setup_reflectivity.py install
-```
-
 ## Windows (using WSL)
 
[…]
