+++ .github/workflows/go-rxr-tests.yml
@@ -25,20 +25,24 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
-        python-version: '3.10.12'
+        python-version: '3.10.14'
 
-    - name: Install dependencies
-      run: |
-        pip install --upgrade pip setuptools
-        pip install -r requirements.txt
+    # - name: Install dependencies
+    #   run: |
+    #     pip install --upgrade pip setuptools
+    #     pip install -r requirements.txt
 
-    - name: Install PyQt5 Dependencies
-      run: |
-        sudo apt-get install -y pyqt5-dev libqt5multimedia5-plugins
+    # - name: Install PyQt5 Dependencies
+    #   run: |
+    #     sudo apt-get install -y pyqt5-dev libqt5multimedia5-plugins
 
-    - name: Install Reflectivity Package
-      run: |
-        python setup_reflectivity.py install
+    # - name: Install Reflectivity Package
+    #   run: |
+    #     python setup_reflectivity.py install
+
+    - name: Install package
+    run: |
+      pip install .
 
     - name: Run tests
       run: |
