+++ .github/workflows/go-rxr-tests.yml
@@ -22,6 +22,11 @@ jobs:
     - name: Check out the code
       uses: actions/checkout@v3
 
+    - name: Move code up one directory level
+      run: |
+        mv ${GITHUB_WORKSPACE}/GO-RXR/* ${GITHUB_WORKSPACE}
+        rmdir ${GITHUB_WORKSPACE}/GO-RXR
+
     - name: Set environment variable for root directory
       run: echo "ROOT_DIR=$(pwd)" >> $GITHUB_ENV
 
@@ -32,31 +37,11 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: '3.10.14'
-
-    # - name: Install dependencies
-    #   run: |
-    #     pip install --upgrade pip setuptools
-    #     pip install -r requirements.txt
-
-    # - name: Install PyQt5 Dependencies
-    #   run: |
-    #     sudo apt-get install -y pyqt5-dev libqt5multimedia5-plugins
-
-    # - name: Install Reflectivity Package
-    #   run: |
-    #     python setup_reflectivity.py install
-
+        
     - name: Install package
       run: |
         pip install .
 
     - name: Run tests
-      # run: |
-        # python -m unittest discover -s TESTS -p "test_data_fitting.py"
-        # python -m unittest discover -s TESTS -p "test_data_structure.py"
-        # python -m unittest discover -s TESTS -p "test_density.py"
-        # python -m unittest discover -s TESTS -p "test_material_model.py"
-        # python -m unittest discover -s TESTS -p "test_material_structure.py"
-        # python -m unittest discover -s TESTS -p "test_reflectivity.py"
       run: |
         python -m unittest discover -s TESTS -p "test_*.py"
