### 2024-09-02 0e08993
Steped back on workflow file to test install dependencies
