+++ .github/workflows/go-rxr-tests.yml
@@ -30,9 +30,23 @@ jobs:
       with:
         python-version: '3.10.14'
 
-    - name: Install package
+    # - name: Install app dependencies
+    #   run: |
+    #     pip install .
+
+    - name: Install dependencies
       run: |
-        pip install .
+        pip install --upgrade pip setuptools
+        pip install -r requirements.txt
+
+    - name: Install PyQt5 Dependencies
+      run: |
+        sudo apt-get install -y pyqt5-dev libqt5multimedia5-plugins
+
+    - name: Install Reflectivity Package
+      run: |
+        python setup_reflectivity.py install
+        
 
     - name: Run tests
       run: |
