### 2024-09-02 d4336b9
Harcoded removing GO-RXR duplicates in root and uses _PATH
