+++ TESTS/test_reflectivity.py
@@ -8,7 +8,7 @@ sys.path.append(parent_dir)
 
 import numpy as np
 import UTILS.material_structure as ms
-from UTILS import TESTS_DIR
+from UTILS import TESTS_PATH
 import unittest
 
 # This test script can be executed by inputting
@@ -27,19 +27,19 @@ class TestReflectivity(unittest.TestCase):
     def __init__(self, methodName: str = "runTest") -> None:
         super().__init__(methodName)
 
-        self.tests_dir = TESTS_DIR
+        self.TESTS_PATH = TESTS_PATH
 
         self.filename1 = 'Si-Al-10A-550.txt'
-        self.my_path1 = self.tests_dir + '/test_data/' + self.filename1 
+        self.my_path1 = self.TESTS_PATH + '/test_data/' + self.filename1 
 
         self.filename2 = 'Si-Al-50A-550.txt'
-        self.my_path2 = self.tests_dir + '/test_data/' + self.filename2
+        self.my_path2 = self.TESTS_PATH + '/test_data/' + self.filename2
 
         self.filename3 = 'AgBr-AlAs-50A-550.txt'
-        self.my_path3 = self.tests_dir + '/test_data/' + self.filename3
+        self.my_path3 = self.TESTS_PATH + '/test_data/' + self.filename3
 
         self.filename4 = 'AgBr-AlAs-50A-550_pi.txt'
-        self.my_path4 = self.tests_dir + '/test_data/' + self.filename4
+        self.my_path4 = self.TESTS_PATH + '/test_data/' + self.filename4
 
 
     def test_single_element_sigma(self):
