### 2024-09-02 6b3e6c2
Update worflow file - all tests explicitely
