+++ .github/workflows/go-rxr-tests.yml
@@ -33,4 +33,9 @@ jobs:
 
     - name: Run tests
       run: |
-        python -m unittest discover -s TESTS -p "test_*.py"
+        python -m unittest discover -s TESTS -p "test_data_fitting.py"
+        python -m unittest discover -s TESTS -p "test_data_structure.py"
+        python -m unittest discover -s TESTS -p "test_density.py"
+        python -m unittest discover -s TESTS -p "test_material_model.py"
+        python -m unittest discover -s TESTS -p "test_material_structure.py"
+        python -m unittest discover -s TESTS -p "test_reflectivity.py
