### 2024-09-02 2b00948
Update tests to be run
