+++ TESTS/test_data_fitting.py
@@ -26,11 +26,13 @@ class TestDataFitting(unittest.TestCase):
     def __init__(self, methodName: str = "runTest") -> None:
         super().__init__(methodName)
 
-        parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
-        self.TESTS_PATH = os.path.join(parent_dir, 'TESTS')
+        # parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+        # self.TESTS_PATH = os.path.join(parent_dir, 'TESTS')
 
-        print('listdir', os.listdir(self.TESTS_PATH))
-        print('listdir', os.listdir(self.TESTS_PATH + '/test_data'))
+        # print('listdir', os.listdir(self.TESTS_PATH))
+        # print('listdir', os.listdir(self.TESTS_PATH + '/test_data'))
+
+        self.TESTS_PATH = TESTS_PATH
 
         self.filename = 'Pim4uc_test.h5'
         self.my_path = self.TESTS_PATH + '/test_data/' + self.filename
