+++ .github/workflows/go-rxr-tests.yml
@@ -38,4 +38,4 @@ jobs:
         python -m unittest discover -s TESTS -p "test_density.py"
         python -m unittest discover -s TESTS -p "test_material_model.py"
         python -m unittest discover -s TESTS -p "test_material_structure.py"
-        python -m unittest discover -s TESTS -p "test_reflectivity.py
+        python -m unittest discover -s TESTS -p "test_reflectivity.py"
