### 2024-09-02 #6577 @jpcurbelo
> thanks for considering the comments.

Quick update. We modified the installation process and now all the dependencies are to be installed by running "pip install .". We also updated the Readme file. This was tested on Ubuntu and we plan to test it - and update the instructions - in a WSL environment as well.
