+++ README.md
@@ -54,36 +54,55 @@ Tested on Ubuntu 22.04
 $ git clone https://github.com/lucaskorol21/GO-RXR.git
 ```
 
-#### 2. Prerequisites (Tested with Python 3.10.14)
+#### 2. Prerequisites and Setup (Tested with Python 3.10.14)
 
-We recommend creating a virtual enviromnment:
+1. **Set Up the Virtual Environment**
 
-```bash
-$ virtualenv -p python3.10 venv-go-rxr
-```
+   - **Create the virtual environment**:  
+     Inside the `GO-RXR` directory:  
+      ```bash
+      $ virtualenv -p python3.10 venv-go-rxr
+      ```
 
-Install the Python libraries using `pip`
+   - **Activate the virtual environment**:  
+     Activate the virtual environment directory:  
+     ```bash
+     (venv-go-rxr) $ source venv-go-rxr/bin/activate
+     ```
 
-```bash
-(venv-go-rxr) $ pip install .
-```
+2. **Install Package Dependencies**
+
+    Install the required Python packages using `pip`:
+
+    ```bash
+    (venv-go-rxr) $ pip install .
+    ```
 
 This command will handle the installation of all dependencies, including compiling any Cython extensions.
 
-If the setup file does not work, then the libraries in the `requirements.txt` file can be installed.
 
-```bash
-(venv-go-rxr) $ pip install -r requirements.txt
-```
 
-#### For `matplotlib`, ensure that you have `Pilow` installed
+### Running the Application
+
+  - Go back to the root of the repository and run the GUI:  
+    ```bash
+    (venv-go-rxr) $ python GUI_GO.py
+    ```
+
+After running the `GUI_GO.py` file, the start screen will be displayed:
+
+![Start screen of GO-RXR. The interface is divided into three main sections: 1) Toolbar, 2) Workspace Navigator, and 3) Workspace Area. \label{fig:start-screen}](FIGURES/go-rxr-start.png)
+
+### Troubleshooting Installation Issues
+
+#### For `matplotlib`, ensure that you have `Pillow` installed
 ```bash
 (venv-go-rxr) $ pip install Pillow
 ```
 
 #### Resolving `PyQt5` Conflicts.
 
-If you encounter conflicts with the PyQt5 package during installation or usage, [this discussion](https://stackoverflow.com/questions/74997556/problem-with-pyqt5-in-ubuntu-22-04-not-fount-zdapvm) might be uselful. Try the following steps:
+If you encounter conflicts with the PyQt5 package during installation or usage, [this discussion](https://stackoverflow.com/questions/74997556/problem-with-pyqt5-in-ubuntu-22-04-not-fount-zdapvm) might be useful. Try the following steps:
 
 ```bash
 # Install necessary dependencies
@@ -110,7 +129,7 @@ This configuration was tested on Windows 11 Home and Windows 11 Education.
    - **Enable WSL**:  
      Open a PowerShell terminal as Administrator and run:  
      ```bash
-     $ wsl --install
+      $ wsl --install
      ```
      Recommended: install Ubuntu 22.04 if available.
 
@@ -126,26 +145,27 @@ This configuration was tested on Windows 11 Home and Windows 11 Education.
    - **Open Ubuntu Terminal**:  
      Update the package list:  
      ```bash
-     $ sudo apt update
+      $ sudo apt update
      ```
      
      Upgrade the installed packages:  
      ```bash
-     $ sudo apt full-upgrade -y
+      $ sudo apt full-upgrade -y
      ```
+
 3. **Set Up Python Environment**
 
    - **Check/Install Python version**:  
      You may already have Python installed with Ubuntu. Check the version:  
      ```bash
-     $ python3 --version
+      $ python3 --version
      ```
[…]
