### 2024-09-10 #7165 @trygvrad
I have some concerns regarding the text, and I think the Statement of Need needs to be reworked.

* The first paragraph describes how data was successfully analysed by this software. This proves that the software is useful (by example), but it does not really describe the need.
* The second paragraph describes challenges, how GO-PXR solves them, and how that is useful. However, the specifics of the challenges are missing, and the challenges are instead described in general terms (multifaceted, light-matter interactions etc.).
* These issues are compounded by the Comparison section, where 'the general problem of RXR data analysis' is referenced, but as a reader it is still not clear to me what these are.
* On line 49 you mention 'element-specific interactions in RXR' as an area which GO-RXR is better than the alternatives. This should refer to the statement of need where this is described as a challenge.

The concrete challenges that this software addresses must be clear, and be at the top of the Statement of Need.

It would help a lot if you could use specific words where you can.
For example, whenever you write 'interaction', see if you can replace it with e.g. 'scattering' or 'absorption'.
This goes for more complex statements as well:
`wide range of materials and conditions` → `oxides, polymers and metals under x-ray energies from ~1 keV to 100 keV`
`diverse challenges` → `absorption edges and resonance peaks`
(Don't use my suggestions here, I just made them up, but you need this level of specificity.)

I noted the following questions as I read the manuscript, which are points where the specificity can be improved to provide more substance to the reader.
* l10 its extreme complexity → what complexity?
* l11 independent variables → what parameters?
* l31 multifaceted challenges → what challenges?
* l32 expertise in materials and the physics of light-matter interactions → what interactions? absorption, scattering, resonant scattering?
* l33 intuition about parameter adjustments → what parameters?
* l37 The software’s capability to capture features in experimental data → what features
* ...
It is not necessary for  you to change all of these, as some of them refer back to each other, but you will need to be more specific overall.
