### 2024-09-20 lucaskorol21/GO-RXR#42 @JohannesBuchner
Installation on Ubuntu Linux, and segfault
Here is my attempt to install following the instructions:

```
⌁ [user:~/Downloads] $ git clone https://github.com/lucaskorol21/GO-RXR
Cloning into 'GO-RXR'...
remote: Enumerating objects: 6804, done.
remote: Total 6804 (delta 525), reused 713 (delta 429), pack-reused 5829 (from 1)
Resolving deltas: 100% (3068/3068), done.
⌁ [user:~/Downloads] $ cd GO-RXR/^C
⌁ [user:~/Downloads] 130 $ virtualenv -p python3.10 venv-go-rxr
RuntimeError: failed to find interpreter for Builtin discover of python_spec='python3.10'
```

I then tried `virtualenv venv-go-rxr`, which uses CPython3.12.3.final.0-64.

Installing worked:
```
 main⌁ [user:~/Downloads/GO-RXR] [venv-go-rxr] ± pip install .
Processing /mnt/data/daten/PostDoc2/home/Downloads/GO-RXR
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting Cython==3.0.8 (from GO-RXR==1.0)
  Downloading Cython-3.0.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (3.2 kB)
Collecting numpy==1.26.3 (from GO-RXR==1.0)
  Downloading numpy-1.26.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (61 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 61.2/61.2 kB 1.9 MB/s eta 0:00:00
Collecting h5py==3.10.0 (from GO-RXR==1.0)
  Downloading h5py-3.10.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (2.5 kB)
Collecting numba==0.59.0 (from GO-RXR==1.0)
  Downloading numba-0.59.0-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (2.7 kB)
Collecting pyqtgraph==0.13.3 (from GO-RXR==1.0)
  Downloading pyqtgraph-0.13.3-py3-none-any.whl.metadata (1.3 kB)
Collecting scipy==1.12.0 (from GO-RXR==1.0)
  Downloading scipy-1.12.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (60 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 60.4/60.4 kB 4.0 MB/s eta 0:00:00
Collecting matplotlib==3.8.1 (from GO-RXR==1.0)
  Downloading matplotlib-3.8.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.8 kB)
Collecting PyQt5==5.15.2 (from GO-RXR==1.0)
  Downloading PyQt5-5.15.2-5.15.2-cp35.cp36.cp37.cp38.cp39-abi3-manylinux2014_x86_64.whl.metadata (2.1 kB)
Collecting contourpy>=1.0.1 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading contourpy-1.3.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.4 kB)
Collecting cycler>=0.10 (from matplotlib==3.8.1->GO-RXR==1.0)
  Using cached cycler-0.12.1-py3-none-any.whl.metadata (3.8 kB)
Collecting fonttools>=4.22.0 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading fonttools-4.53.1-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (162 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 162.6/162.6 kB 5.5 MB/s eta 0:00:00
Collecting kiwisolver>=1.3.1 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading kiwisolver-1.4.7-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (6.3 kB)
Collecting packaging>=20.0 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading packaging-24.1-py3-none-any.whl.metadata (3.2 kB)
Collecting pillow>=8 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading pillow-10.4.0-cp312-cp312-manylinux_2_28_x86_64.whl.metadata (9.2 kB)
Collecting pyparsing>=2.3.1 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading pyparsing-3.1.4-py3-none-any.whl.metadata (5.1 kB)
Collecting python-dateutil>=2.7 (from matplotlib==3.8.1->GO-RXR==1.0)
  Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl.metadata (8.4 kB)
Collecting llvmlite<0.43,>=0.42.0dev0 (from numba==0.59.0->GO-RXR==1.0)
  Using cached llvmlite-0.42.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (4.8 kB)
Collecting PyQt5-sip<13,>=12.8 (from PyQt5==5.15.2->GO-RXR==1.0)
  Downloading PyQt5_sip-12.15.0-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.whl.metadata (421 bytes)
Collecting six>=1.5 (from python-dateutil>=2.7->matplotlib==3.8.1->GO-RXR==1.0)
  Downloading six-1.16.0-py2.py3-none-any.whl.metadata (1.8 kB)
Downloading Cython-3.0.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (3.5 MB)
Downloading h5py-3.10.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (4.7 MB)
Downloading matplotlib-3.8.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (11.6 MB)
Downloading numba-0.59.0-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl (3.7 MB)
Downloading numpy-1.26.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (18.0 MB)
Downloading PyQt5-5.15.2-5.15.2-cp35.cp36.cp37.cp38.cp39-abi3-manylinux2014_x86_64.whl (68.3 MB)
Downloading pyqtgraph-0.13.3-py3-none-any.whl (960 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 961.0/961.0 kB 6.8 MB/s eta 0:00:00
Downloading scipy-1.12.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (37.8 MB)
Downloading contourpy-1.3.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (320 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 320.5/320.5 kB 6.7 MB/s eta 0:00:00
Using cached cycler-0.12.1-py3-none-any.whl (8.3 kB)
Downloading fonttools-4.53.1-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl (4.9 MB)
Downloading kiwisolver-1.4.7-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (1.5 MB)
Using cached llvmlite-0.42.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (43.8 MB)
Downloading packaging-24.1-py3-none-any.whl (53 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 54.0/54.0 kB 4.3 MB/s eta 0:00:00
Downloading pillow-10.4.0-cp312-cp312-manylinux_2_28_x86_64.whl (4.5 MB)
Downloading pyparsing-3.1.4-py3-none-any.whl (104 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 104.1/104.1 kB 5.9 MB/s eta 0:00:00
Downloading PyQt5_sip-12.15.0-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.whl (281 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 281.4/281.4 kB 6.6 MB/s eta 0:00:00
Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl (229 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 229.9/229.9 kB 6.1 MB/s eta 0:00:00
Using cached six-1.16.0-py2.py3-none-any.whl (11 kB)
Building wheels for collected packages: GO-RXR
  Building wheel for GO-RXR (pyproject.toml) ... done
  Created wheel for GO-RXR: filename=GO_RXR-1.0-cp312-cp312-linux_x86_64.whl size=1286425 sha256=17cf33fe0b3560f2894606c306ce018763b67bb0c77133d58213237d52929192
  Stored in directory: /home/user/.cache/pip/wheels/75/36/cd/ef26e8a7b325eed21353f9b6432b7750316b29c0332995dcdf
Successfully built GO-RXR
Installing collected packages: six, PyQt5-sip, pyparsing, pillow, packaging, numpy, llvmlite, kiwisolver, fonttools, Cython, cycler, scipy, python-dateutil, pyqtgraph, PyQt5, numba, h5py, contourpy, matplotlib, GO-RXR
Successfully installed Cython-3.0.8 GO-RXR-1.0 PyQt5-5.15.2 PyQt5-sip-12.15.0 contourpy-1.3.0 cycler-0.12.1 fonttools-4.53.1 h5py-3.10.0 kiwisolver-1.4.7 llvmlite-0.42.0 matplotlib-3.8.1 numba-0.59.0 numpy-1.26.3 packaging-24.1 pillow-10.4.0 pyparsing-3.1.4 pyqtgraph-0.13.3 python-dateutil-2.9.0.post0 scipy-1.12.0 six-1.16.0

[notice] A new release of pip is available: 24.0 -> 24.2
[notice] To update, run: pip install --upgrade pip
```

But running fails with a segfault:
```
$ python3 GUI_GO.py
Segmentation fault (core dumped)
```

GDB says:
```
gdb python3
GNU gdb (Ubuntu 15.0.50.20240403-0ubuntu1) 15.0.50.20240403-git
Copyright (C) 2024 Free Software Foundation, Inc.
License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>
This is free software: you are free to change and redistribute it.
There is NO WARRANTY, to the extent permitted by law.
Type "show copying" and "show warranty" for details.
This GDB was configured as "x86_64-linux-gnu".
Type "show configuration" for configuration details.
For bug reporting instructions, please see:
<https://www.gnu.org/software/gdb/bugs/>.
Find the GDB manual and other documentation resources online at:
    <http://www.gnu.org/software/gdb/documentation/>.

For help, type "help".
Type "apropos word" to search for commands related to "word"...
Reading symbols from python3...
Downloading separate debug info for /usr/bin/python3.12
Reading symbols from /home/user/.cache/debuginfod_client/a36ec6809086f2d7b24ed7ef83d3de5555ffbe8f/debuginfo...
(gdb) run GUI_GO.py
Starting program: /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/bin/python3 GUI_GO.py
Downloading separate debug info for system-supplied DSO at 0x7ffff7fc3000
Downloading separate debug info for /lib/x86_64-linux-gnu/libz.so.1
Downloading separate debug info for /lib/x86_64-linux-gnu/libexpat.so.1
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/core/_multiarray_umath.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/core/../../numpy.libs/libopenblas64_p-r0-0cf96a72.3.23.dev.so
Downloading separate debug info for /lib/x86_64-linux-gnu/libgcc_s.so.1
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/core/../../numpy.libs/libgfortran-040039e1.so.5.0.0
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/core/../../numpy.libs/libquadmath-96973f99.so.0.0.0
[New Thread 0x7ffff2e006c0 (LWP 378516)]
[New Thread 0x7ffff04006c0 (LWP 378517)]
[New Thread 0x7fffefa006c0 (LWP 378518)]
[New Thread 0x7fffed0006c0 (LWP 378519)]
[New Thread 0x7fffea6006c0 (LWP 378520)]
[New Thread 0x7fffe5c006c0 (LWP 378521)]
[New Thread 0x7fffe32006c0 (LWP 378522)]
[New Thread 0x7fffe28006c0 (LWP 378523)]
[New Thread 0x7fffdde006c0 (LWP 378524)]
[New Thread 0x7fffdd4006c0 (LWP 378525)]
[New Thread 0x7fffd8a006c0 (LWP 378526)]
[New Thread 0x7fffd60006c0 (LWP 378527)]
[New Thread 0x7fffd56006c0 (LWP 378528)]
[New Thread 0x7fffd2c006c0 (LWP 378529)]
[New Thread 0x7fffd02006c0 (LWP 378530)]
[New Thread 0x7fffcb8006c0 (LWP 378531)]
[New Thread 0x7fffc8e006c0 (LWP 378532)]
[New Thread 0x7fffc64006c0 (LWP 378533)]
[New Thread 0x7fffc5a006c0 (LWP 378534)]
Downloading separate debug info for /usr/lib/python3.12/lib-dynload/_contextvars.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/core/_multiarray_tests.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /usr/lib/python3.12/lib-dynload/_ctypes.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /lib/x86_64-linux-gnu/libffi.so.8
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/linalg/_umath_linalg.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/fft/_pocketfft_internal.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/mtrand.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/bit_generator.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_common.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /usr/lib/python3.12/lib-dynload/_hashlib.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /lib/x86_64-linux-gnu/libcrypto.so.3
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_bounded_integers.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_mt19937.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_philox.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_pcg64.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_sfc64.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/numpy/random/_generator.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/_lib/_ccallback_c.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/_ufuncs.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/../../scipy.libs/libopenblasp-r0-23e5df77.3.21.dev.so
[New Thread 0x7fffbee006c0 (LWP 378539)]
[New Thread 0x7fffbe4006c0 (LWP 378540)]
[New Thread 0x7fffbba006c0 (LWP 378541)]
[New Thread 0x7fffb90006c0 (LWP 378542)]
[New Thread 0x7fffb66006c0 (LWP 378543)]
[New Thread 0x7fffb3c006c0 (LWP 378544)]
[New Thread 0x7fffb12006c0 (LWP 378545)]
[New Thread 0x7fffae8006c0 (LWP 378546)]
[New Thread 0x7fffabe006c0 (LWP 378547)]
[New Thread 0x7fffa74006c0 (LWP 378548)]
[New Thread 0x7fffa4a006c0 (LWP 378549)]
[New Thread 0x7fffa40006c0 (LWP 378550)]
[New Thread 0x7fffa16006c0 (LWP 378551)]
[New Thread 0x7fff9cc006c0 (LWP 378552)]
[New Thread 0x7fff9c2006c0 (LWP 378553)]
[New Thread 0x7fff998006c0 (LWP 378554)]
[New Thread 0x7fff96e006c0 (LWP 378555)]
[New Thread 0x7fff924006c0 (LWP 378556)]
[New Thread 0x7fff8fa006c0 (LWP 378557)]
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/_ufuncs_cxx.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /lib/x86_64-linux-gnu/libstdc++.so.6
Downloading separate debug info for /usr/lib/python3.12/lib-dynload/_bz2.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /lib/x86_64-linux-gnu/libbz2.so.1.0
Downloading separate debug info for /usr/lib/python3.12/lib-dynload/_lzma.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /lib/x86_64-linux-gnu/liblzma.so.5
[Detaching after vfork from child process 378563]
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/_specfun.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/_comb.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_fblas.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_flapack.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_cythonized_array_utils.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/cython_lapack.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_solve_toeplitz.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_flinalg.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_decomp_lu_cython.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_matfuncs_sqrtm_triu.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_matfuncs_expm.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/cython_blas.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/linalg/_decomp_update.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/_sparsetools.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/_csparsetools.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/linalg/_dsolve/_superlu.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/linalg/_eigen/arpack/_arpack.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_shortest_path.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_tools.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_traversal.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_min_spanning_tree.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_flow.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_matching.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/sparse/csgraph/_reordering.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/special/_ellip_harm_2.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/interpolate/_fitpack.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/interpolate/dfitpack.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/optimize/_minpack2.cpython-312-x86_64-linux-gnu.so
Downloading separate debug info for /mnt/data/daten/PostDoc2/home/Downloads/venv-go-rxr/lib/python3.12/site-packages/scipy/optimize/_group_columns.cpython-312-x86_64-linux-gnu.so
[…]
