### 2024-09-20 lucaskorol21/GO-RXR#42 @JohannesBuchner
I was able to resolve the issue by uninstalling pyqt5 and reinstalling it.
