### 2024-09-20 lucaskorol21/GO-RXR#42 @JohannesBuchner
Reinstalling in a fresh virtual environment following the instructions as above gives the same segfault though. Unpinning the PyQT5 dependency helped.
