### 2024-09-20 lucaskorol21/GO-RXR#43 @JohannesBuchner
Help menu buttons don't work
License:
```
Traceback (most recent call last):
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/GUI_GO.py", line 9256, in _license
    license = licenseWidget()
              ^^^^^^^^^^^^^^^
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/GUI_GO.py", line 11487, in __init__
    with open('license.txt', 'r',encoding='ISO-8859-1') as f:
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: 'license.txt'
```
About:
`gio: file:GO-RXR_UserGuide_v0.3.1.pdf: Operation not supported`

Tips:
```
Traceback (most recent call last):
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/GUI_GO.py", line 9264, in _tips
    license = tipsWidget()
              ^^^^^^^^^^^^
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/GUI_GO.py", line 11519, in __init__
    with open('tips.txt', 'r',encoding='ISO-8859-1') as f:
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: 'tips.txt'

```
