### 2024-09-20 lucaskorol21/GO-RXR#44 @JohannesBuchner
Saving simulation gives error
```
Traceback (most recent call last):
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/GUI_GO.py", line 9002, in _saveSimulation
    ds.saveSimulationHDF5(self.fname, sim_dict, self.version)
  File "/mnt/data/daten/PostDoc2/home/Downloads/GO-RXR/UTILS/data_structure.py", line 492, in saveSimulationHDF5
    simulated = f['Simulated_data']
                ~^^^^^^^^^^^^^^^^^^
  File "h5py/_debian_h5py_serial/_objects.pyx", line 54, in h5py._debian_h5py_serial._objects.with_phil.wrapper
  File "h5py/_debian_h5py_serial/_objects.pyx", line 55, in h5py._debian_h5py_serial._objects.with_phil.wrapper
  File "/usr/lib/python3/dist-packages/h5py/_debian_h5py_serial/_hl/group.py", line 357, in __getitem__
    oid = h5o.open(self.id, self._e(name), lapl=self._lapl)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "h5py/_debian_h5py_serial/_objects.pyx", line 54, in h5py._debian_h5py_serial._objects.with_phil.wrapper
  File "h5py/_debian_h5py_serial/_objects.pyx", line 55, in h5py._debian_h5py_serial._objects.with_phil.wrapper
  File "h5py/_debian_h5py_serial/h5o.pyx", line 189, in h5py._debian_h5py_serial.h5o.open
KeyError: "Unable to open object (object 'Simulated_data' doesn't exist)"
```
