### 2024-09-23 #7165 @jpcurbelo
> I followed the install, and then Example 1 in the tutorial until "Update Sample", and then clicked around a bit. Please see the issues I opened [lucaskorol21/GO-RXR#42](https://github.com/lucaskorol21/GO-RXR/issues/42) [lucaskorol21/GO-RXR#43](https://github.com/lucaskorol21/GO-RXR/issues/43) [lucaskorol21/GO-RXR#44](https://github.com/lucaskorol21/GO-RXR/issues/44)
>
> Additionally, here are some comments:
>
>     * LICENCE file: you may want to use legal name of the code authors.
>
>
> INSTALL instructions:
>
>     * you could make entering the repo directory (cd) more prominent / explicit
>
>     * the dependencies are all pinned, which is good practice for GUI applications. But it caused install issues for me. [Installation on Ubuntu Linux, and segfault lucaskorol21/GO-RXR#42](https://github.com/lucaskorol21/GO-RXR/issues/42)
>
>
> Running pip install commands is dodgy, because you don't know the exact computing environment and what damage this can cause:
>
> ```
> try:
>     import numpy
> except ImportError:
>     print("Numpy is not installed. Installing now...")
>     subprocess.check_call([sys.executable, "-m", "pip", "install", "numpy==1.26.3"])
>     import numpy
>
> try:
>     from setuptools import setup, find_packages, Extension
> except ImportError:
>     print("Setuptools is not installed. Installing now...")
>     subprocess.check_call([sys.executable, "-m", "pip", "install", "--upgrade", "pip", "setuptools"])
>     from setuptools import setup, find_packages, Extension
>
> try:
>     from Cython.Build import cythonize
>     from Cython.Distutils import build_ext
> except ImportError:
>     print("Cython is not installed. Installing now...")
>     subprocess.check_call([sys.executable, "-m", "pip", "install", "Cython==3.0.8"])
>     from Cython.Build import cythonize
>     from Cython.Distutils import build_ext
> ```
>
> It would be better to have a new section before the install where you describe how to install the prerequesites.
>
> In fact, setup.py is altogether discouraged now, in favor of non-procedural install instructions (pyproject.toml). This also solves previous cython issues.
>
> Documentation:
>
>     * Is chi² and L2 norm not the same thing?
>
>
> Taking a break now.

Hi @JohannesBuchner. We appreciate your comments and attention to details. I'll be answering the issues you opened shortly.
