### 2024-09-24 lucaskorol21/GO-RXR#42 @jpcurbelo
Hi @JohannesBuchner, thank you for your feedback! I'll try to address your points here:

1- **Python Version Requirement**: To run `virtualenv -p python3.10 venv-go-rxr`,`Python 3.10` needs to be installed on the machine. We should have made that explicit in the README file and I will include it in the next update. I’ll add the following step (with more details) to the guide:
```
sudo apt update
sudo apt install python3.10 python3.10-venv python3.10-dev
```
2- **Segmentation Fault Issue with Python 3.12**: I was able to reproduce the issue after creating a `virtualenv` with `Python 3.12.3`. I will emphasize that `Python 3.10` is the recommended version in the README. Additionally, I’ll include the troubleshooting steps you followed and will suggest unpinning  the PyQT5 dependency in a dedicated troubleshooting section.
