### 2024-09-24 0fd50b0
Update readme: instructions for python3.10 and env
