+++ PAPER/paper.md
@@ -37,15 +37,20 @@ This paper presents an overview of GO-RXR, highlighting its functionality, examp
 
 # Statement of Need
 
-One of the challenges addressed by [@korol_MSc_2023] included the analysis of LaMnO<sub>3</sub>/SrTiO<sub>3</sub> thin-film heterostructures. The data for these samples were collected at the resonant elastic and inelastic x-ray scattering beamline (REIXS) at the Candian Light Source (CLS) in 2017. Although the initial data collection took only three days, attempts to analyze it with the tools available in 2021 yielded little success. In 2023, use of the GO-RXR tool led to a successful analysis, highlighting the efficacy of the software in overcoming longstanding analysis barriers.
+RXR offers unique insights into the depth-dependent crystal, electronic, and magnetic structures of quantum materials, enabling the investigation of nanoscale characteristics of new candidate materials with a precision unmatched by any other current experimental technique **[REF - new?]**. Despite its potential, RXR remains significantly underutilized, with far fewer publications compared to techniques such as x-ray absorption spectroscopy. The main challenge hindering the widespread adoption of RXR lies in the extreme difficulty of data analysis, which requires both large-scale computational quantum mechanics simulations and the fitting of reflectivity models with numerous independent variables, such as layer thickness, interfacial roughness, and complex refractive indices that vary with energy. Each of these parameters must be finely adjusted to match the experimental data, making the process intricate and labor-intensive. For example, the oscillations in the Kiessig fringes during theta/two-theta reflectivity scans are directly related to the thickness of the film, while the decay of these fringes indicates the roughness of various interfaces within the material. This makes the analysis process for each sample highly demanding and time-consuming. Consequently, experimental advancements have far outpaced the progress in analytical methods, leaving a substantial amount of collected data unexplored **[REF???]**.
 
-The analysis of RXR presents multifaceted challenges, extending beyond computational aspects to encompass expertise in materials and the physics of light-matter interactions. This expertise is pivotal because it provides intuition about parameter adjustments and guides the direction of data analysis to achieve desired outcomes. In addressing this challenge, GO-RXR integrates global optimization algorithms, thereby lowering the expertise threshold necessary for effective data analysis. Through the extensive development of GO-RXR, diverse global optimization algorithms and unique objective functions were thoroughly explored. The software's capability to capture features in experimental data without exhaustive parameter understanding significantly reduces the expertise required. GO-RXR serves as a valuable scientific tool for material scientists, offering advanced capabilities to streamline data analysis and reduce the expertise barrier, ultimately facilitating breakthrough discoveries in the field of materials science.
[…]
