### 2024-09-25 aa4928a
Fix help menu. Add files and pdf call
