### 2024-09-25 lucaskorol21/GO-RXR#43 @jpcurbelo
We just updated the Help Menu. Now the options listed are:
- User Guide (loads the pdf file with the documentation)
- Tips (load a list of useful tips)
- License (loads the updated license file - we included legal names here)
