### 2024-09-25 lucaskorol21/GO-RXR#44 @jpcurbelo
Hi! I wasn't able to reproduce exactly this issue, although, got `KeyError: "Unable to synchronously open object (object 'Simulated_data' doesn't exist)"` when saving the simulation. However, a `demo.h5` file is saved in the root dir as expected. We are working on fixing this and improving the saving process. In the meantime, could you, please, let us know if any file was saved when you tried? Thanks.
