+++ pyproject.toml
@@ -1,7 +1,38 @@
 [build-system]
-requires = [
-        "setuptools",
-        "wheel",
-        "cython",
-        "numpy<2",
+requires = ["setuptools>=61.0", "wheel", "cython>=3.0.0", "numpy>=1.26"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "GO-RXR"
+version = "1.0"
+description = "GO-RXR: Global Optimization of Resonant X-ray Reflectometry Tool for Analyzing Quantum Materials"
+readme = "README.md"
+license = {text = "BSD 3-Clause License"}
+authors = [
+    {name = "Lucas Korol", email = "lsk601@usask.ca"},
+    {name = "Robert J. Green", email = "robert.green@usask.ca"}
+    {name = "Jesus P. Curbelo", email = "jesus.curbelo@usask.ca"}
+    {name = "Raymond J. Spiteri", email = "spiteri@cs.usask.ca"}
 ]
+requires-python = ">=3.10"
+
+dependencies = [
+    "numpy>=1.26",
+    "Cython>=3.0.0",
+    "h5py>=3.10.0",
+    "numba>=0.59.0",
+    "pyqtgraph>=0.13.0",
+    "scipy>=1.12.0",
+    "matplotlib>=3.8.0",
+    "PyQt5>=5.15.0"
+]
+
+[project.urls]
+homepage = "https://github.com/lucaskorol21/GO-RXR"
+
+[tool.setuptools]
+packages = ["Magnetic_Scattering_Factor", "Scattering_Factor"]
+include-package-data = true
+
+[tool.setuptools.package-data]
+"Magnetic_Scattering_Factor" = ["*.md"]
