### 2024-10-22 2070a70
Add commas to author list in toml
