### 2024-10-23 d11e90b
Add Spiteri's edits to paper
