+++ PAPER/paper.md
@@ -31,22 +31,22 @@ bibliography: paper.bib
 
 # Summary
 
-Resonant x-ray reflectometry (RXR) is a cutting-edge synchrotron technique used to characterize the depth-dependent structure of quantum materials [@keimer_moore_NPh_2017; @green-etal_SRN_2020]. However, the main challenge impeding the success of RXR data analysis lies in its extreme complexity, driven by complicated model construction and the fitting of numerous independent variables. This complexity results in prolonged analysis periods that demand significant engagement from researchers. In response to these challenges, the Global Optimization of Resonant X-ray Reflectometry (GO-RXR) software emerged from rigorous development efforts as a main contribution from the work by [@korol_MSc_2023]. GO-RXR streamlines data analysis, enhances visualization, and reduces the expertise required, offering researchers a more efficient means to analyze RXR data. 
+Resonant x-ray reflectometry (RXR) is a cutting-edge synchrotron technique used to characterize the depth-dependent structure of quantum materials [@keimer_moore_NPh_2017; @green-etal_SRN_2020]. However, the main challenge impeding the success of RXR data analysis lies in the complexity of the workflow, driven by complicated model construction and the fitting of numerous parameters. This workflow complexity results in prolonged analysis periods that demand significant engagement from researchers. In response to these challenges, the Global Optimization of Resonant X-ray Reflectometry (GO-RXR) software emerged from rigorous development efforts as a main contribution from the work in [@korol_MSc_2023]. GO-RXR streamlines data analysis, enhances visualization, and reduces the specialized expertise required, offering researchers a more efficient means to analyze RXR data.
 
 This paper presents an overview of GO-RXR, highlighting its functionality, example use-cases, and impact in materials science research. Through its comprehensive approach and user-friendly design, GO-RXR offers researchers an efficient tool for analyzing RXR data, facilitating breakthroughs in understanding complex material systems. Additionally, publications and ongoing research utilizing GO-RXR underscore its versatility and impact in advancing scientific exploration.
 
 # Statement of Need
 
-RXR offers unique insights into the depth-dependent crystal, electronic, and magnetic structures of quantum materials, enabling the investigation of nanoscale characteristics of new candidate materials with a precision unmatched by any other current experimental technique **[REF - new?]**. Despite its potential, RXR remains significantly underutilized, with far fewer publications compared to techniques such as x-ray absorption spectroscopy. The main challenge hindering the widespread adoption of RXR lies in the extreme difficulty of data analysis, which requires both large-scale computational quantum mechanics simulations and the fitting of reflectivity models with numerous independent variables, such as layer thickness, interfacial roughness, and complex refractive indices that vary with energy. Each of these parameters must be finely adjusted to match the experimental data, making the process intricate and labor-intensive. For example, the oscillations in the Kiessig fringes during theta/two-theta reflectivity scans are directly related to the thickness of the film, while the decay of these fringes indicates the roughness of various interfaces within the material. This makes the analysis process for each sample highly demanding and time-consuming. Consequently, experimental advancements have far outpaced the progress in analytical methods, leaving a substantial amount of collected data unexplored **[REF???]**.
[…]
