+++ PAPER/paper.bib
@@ -152,5 +152,31 @@
     url={https://doi.org/10.1107/S1600576719016789},
 }
 
+@article{borrero_etal_2017,
+    title={Site-selective spectroscopy with depth resolution using resonant x-ray reflectometry},
+    author={{Hamann-Borrero}, J. E. and Macke, S. and Gray, B. and et al.},
+    journal={Scientific Reports},
+    volume={7},
+    pages={13792},
+    year={2017},
+    doi={10.1038/s41598-017-12642-7},
+    url={https://doi.org/10.1038/s41598-017-12642-7}
+}
+
+@article{fursich_etal_2018,
+  title = {Theory-restricted resonant x-ray reflectometry of quantum materials},
+  author = {F\"ursich, Katrin and Zabolotnyy, Volodymyr B. and Schierle, Enrico and Dudy, Lenart and Kirilmaz, Ozan and Sing, Michael and Claessen, Ralph and Green, Robert J. and Haverkort, Maurits W. and Hinkov, Vladimir},
+  journal = {Phys. Rev. B},
+  volume = {97},
+  issue = {16},
+  pages = {165126},
+  numpages = {6},
+  year = {2018},
+  month = {Apr},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.97.165126},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.97.165126}
+}
+
 
 
