+++ PAPER/paper.md
@@ -37,7 +37,7 @@ This paper presents an overview of GO-RXR, highlighting its functionality, examp
 
 # Statement of Need
 
-RXR offers unique insights into the depth-dependent crystal, electronic, and magnetic structures of quantum materials, enabling the investigation of nanoscale characteristics of new candidate materials with a precision unmatched by any other current experimental technique [@borrero_etal_2017;fursich_etal_2018; vanderMinne_etal_2023]. Despite its potential, RXR remains significantly underutilized, with far fewer publications compared to techniques such as x-ray absorption spectroscopy. The main challenge hindering the widespread adoption of RXR lies in the difficulty of data analysis, which requires both large-scale computational quantum mechanics simulations and the fitting of reflectivity models with numerous parameters, such as layer thickness, interfacial roughness, and complex refractive indices, that vary with energy. Each of these parameters must be finely adjusted to match experimental data, making the process intricate and labor-intensive. For example, the oscillations in the Kiessig fringes during theta/two-theta reflectivity scans are directly related to the thickness of the film, while the decay of these fringes indicates the roughness of various interfaces within the material. This makes the analysis process for each sample highly demanding and time-consuming. Consequently, experimental advancements have far outpaced the progress in analytical methods, leaving a substantial amount of collected data unexplored.
+RXR offers unique insights into the depth-dependent crystal, electronic, and magnetic structures of quantum materials, enabling the investigation of nanoscale characteristics of new candidate materials with a precision unmatched by any other current experimental technique [@borrero_etal_2017;@fursich_etal_2018; @vanderMinne_etal_2023]. Despite its potential, RXR remains significantly underutilized, with far fewer publications compared to techniques such as x-ray absorption spectroscopy. The main challenge hindering the widespread adoption of RXR lies in the difficulty of data analysis, which requires both large-scale computational quantum mechanics simulations and the fitting of reflectivity models with numerous parameters, such as layer thickness, interfacial roughness, and complex refractive indices, that vary with energy. Each of these parameters must be finely adjusted to match experimental data, making the process intricate and labor-intensive. For example, the oscillations in the Kiessig fringes during theta/two-theta reflectivity scans are directly related to the thickness of the film, while the decay of these fringes indicates the roughness of various interfaces within the material. This makes the analysis process for each sample highly demanding and time-consuming. Consequently, experimental advancements have far outpaced the progress in analytical methods, leaving a substantial amount of collected data unexplored.
 
[…]
