### 2024-10-29 #7165 @jpcurbelo
Hi @zhubonan

We have just pushed the latest updates to both the manuscript and the code. Below, please find our responses to the specific requests:

Manuscript:

We have revised the Statement of Need and Comparison sections in line with @trygvrad's suggestions.

- The updated Statement of Need now provides a clearer explanation of the specific challenges addressed by the software, explicitly describing the necessity for this tool. We have also added references to support these claims.
- We have also included details about the type of "element-specific interactions" handled by the technique and the tool, the factors contributing to the task's complexity, the critical parameters that need to be considered, and the specific features the code extracts from experimental data.
-
Additionally, we have rephrased some terminology for clarity, such as replacing "independent variables" with "parameters" to improve readability.
