### 2024-10-30 #7165 @jpcurbelo
Code:

- We addressed [lucaskorol21/GO-RXR#42](https://github.com/lucaskorol21/GO-RXR/issues/42), providing comments on the issue and updating the README file with recommendations and detailed steps for ensuring the correct version is installed.
- For [lucaskorol21/GO-RXR/issues/43](https://github.com/lucaskorol21/GO-RXR/issues/43), we have updated the Help Menu.
- We have added the legal names and emails of the code authors to the License file for clarity.
- We removed the hard-coded versions from the installation process and updated the steps to use a pyproject.toml file as suggested, improving the installation flexibility.
- Although chi² and the L2 norm are related, they are not the same. chi² is divided by the absolute simulated value, making it a "relative" version of the L2 norm.

To-do:

[lucaskorol21/GO-RXR/issues/44](https://github.com/lucaskorol21/GO-RXR/issues/44): As noted by the reviewer, there are inconsistencies in the load/save buttons and commands. We are actively working on resolving this and expect to have it fixed by next week.
