+++ GUI_GO.py
@@ -6513,6 +6513,8 @@ class GlobalOptimizationWidget(QWidget):
         self.runButton.setStyleSheet('background: green')
         self.runButton.blockSignals(False)
 
+        # Enable Save Summary Button
+
 
     def _run_global_optimization(self):
         """
@@ -8375,11 +8377,11 @@ class ReflectometryApp(QMainWindow):
         self.loadSample.triggered.connect(self._loadSample)
         fileMenu.addAction(self.loadSample)
 
-        # load a file
-        self.loadOrbitals = QAction("&Load Orbitals", self)
-        self.loadOrbitals.triggered.connect(self._loadOrbitals)
-        fileMenu.addAction(self.loadOrbitals)
-        fileMenu.addSeparator()
+        # # load a file
+        # self.loadOrbitals = QAction("&Load Orbitals", self)
+        # self.loadOrbitals.triggered.connect(self._loadOrbitals)
+        # fileMenu.addAction(self.loadOrbitals)
+        # fileMenu.addSeparator()
 
         # save current working file
         self.saveFile = QAction("&Save Workspace", self)
@@ -8396,9 +8398,9 @@ class ReflectometryApp(QMainWindow):
         self.saveSampleFile.triggered.connect(self._saveSample)
         fileMenu.addAction(self.saveSampleFile)
 
-        self.saveOrbitals = QAction("&Save Orbitals", self)
-        self.saveOrbitals.triggered.connect(self._saveOrbitals)
-        fileMenu.addAction(self.saveOrbitals)
+        # self.saveOrbitals = QAction("&Save Orbitals", self)
+        # self.saveOrbitals.triggered.connect(self._saveOrbitals)
+        # fileMenu.addAction(self.saveOrbitals)
 
         # save the simulation
         self.saveSimulationFile = QAction("&Save Simulation", self)
@@ -8464,56 +8466,58 @@ class ReflectometryApp(QMainWindow):
         helpMenu.addAction(self.license)
 
 
-    def _loadOrbitals(self):
-        """
-        Purpose: Load orbital file
-        """
-        fname, _ = QFileDialog.getOpenFileName(self, 'Open File')  # retrieve file name
+    ##### This should be implemented/updated on Jordan's update - summer 2024
+    # def _loadOrbitals(self):
+    #     """
+    #     Purpose: Load orbital file
+    #     """
+    #     fname, _ = QFileDialog.getOpenFileName(self, 'Open File')  # retrieve file name
 
-        #fname = self.fname  # used to check file type
-        self._sampleWidget.sf_dict = copy.deepcopy(mm.retrieve_ff())
+    #     #fname = self.fname  # used to check file type
+    #     self._sampleWidget.sf_dict = copy.deepcopy(mm.retrieve_ff())
 
-        if fname.endswith('.pkl'):
-            import pickle
+    #     if fname.endswith('.pkl'):
+    #         import pickle
 
-            # Load the pickled dictionary from the file
-            with open(fname, 'rb') as file:
-                self._sampleWidget.orbitals = pickle.load(file)
+    #         # Load the pickled dictionary from the file
+    #         with open(fname, 'rb') as file:
+    #             self._sampleWidget.orbitals = pickle.load(file)
 
-        else:
-            messageBox = QMessageBox()
-            messageBox.setWindowTitle("Invalid file name")
-            messageBox.setText("Selected file name or path is not valid. Please select a valid file name.")
-            messageBox.exec()
+    #     else:
+    #         messageBox = QMessageBox()
+    #         messageBox.setWindowTitle("Invalid file name")
+    #         messageBox.setText("Selected file name or path is not valid. Please select a valid file name.")
[…]
