### 2024-11-06 16c5b3c
Save Summary enable after optimization
