### 2024-11-19 ce565c8
Updated saveSample/Sim/Summary with msg
