### 2024-11-19 lucaskorol21/GO-RXR#42 @jpcurbelo
As requested by @JohannesBuchner, we removed the hard-coded versions from the installation process and updated the steps to use a `pyproject.toml` file as suggested, improving the installation flexibility.
