### 2024-11-19 lucaskorol21/GO-RXR#43 @jpcurbelo
Please, let me know if we can close this issue.
