### 2024-11-19 lucaskorol21/GO-RXR#44 @jpcurbelo
We have implemented some updates to the _load/save_ menus:

1- Got rid of the _Load/Save Orbitals_ entries. Commented the functions `_loadOrbitals` and `_saveOrbitals` and added this comment: "This should be implemented/updated on Jordan's update - summer 2024".
2- _Save_ _Simulation/Sample/Summary_ are disable when the GUI is loaded. _Save Simulation_ is also enabled after _Import Dataset_ or _Load ReMagX_.
- _Save Summary_ is  enabled at the end of `optimizationFinished`.
- _Save Summary_ is disabled after the user loads workspace (`_loadFile`) or load sample (`_loadSample`)
-  _Save Simulation_ and _Save Sample_ are enable when the user _Load Workspace_ (`_loadFile`) or _Save Workspace As_ (`_saveAsFile`). This procedure eliminates the issue the reviewer noticed - error when trying to save the simulation before saving the workspace.
- We added a (2s) pop-up message (and terminal log) to let the user know how the _saving_ went.
