### 2024-11-19 #7165 @jpcurbelo
Hi @zhubonan.

Issues [lucaskorol21/GO-RXR#42](https://github.com/lucaskorol21/GO-RXR/issues/42) and [lucaskorol21/GO-RXR#43](https://github.com/lucaskorol21/GO-RXR/issues/43) should be ready to be closed. Also, we just committed the latest updates related to [lucaskorol21/GO-RXR#44](https://github.com/lucaskorol21/GO-RXR/issues/44) and commented all the modifications we implemented.

We believe that we have addressed all the reviewer comments by now. Please, let us know about the next steps to follow. Thanks!
