### 2024-11-20 #7165 @trygvrad
Thank you to the authors for updating the manuscript, it reads much better now.
On line 92 the text reads `and experimental conditions [Examples here]`, which needs to be addressed, but I otherwise have no outstanding issues with the manuscript.

Installation using the updated instructions worked for me on Ubuntu 22.04, and I was able to run the tutorial `DOCS/Tutorial/Tutorial.pdf` without issues.

I was also able to run this on Ubuntu 24.04 after adding the deadsnakes ppa  in order to install python3.10
```
sudo add-apt-repository ppa:deadsnakes/ppa
```
Note that I made the virtual environment with `python3.10 -m venv venv-go-rxr` rather than `virtualenv -p python3.10 venv-go-rxr`. The latter option requires the `python3-virtualenv` package (on Ubuntu), while the former does not.

The instructions did not work in wsl on my arm64 windows laptop, as there was an error installing pyqt5 (below).
I managed to install pyqt6, and update `GUI_GO.py` accordingly, but that produced a different error (further below).
This error appears to relate to the instantiating the graphics, which is somewhat of a complication in wsl.

In any case wsl on an arm64 system is somewhat of a niche use case, and I do not have an amd64 windows system where I can test the software, thus I do not know if this experience is representative.

I hypothesize that updating to python3.12 and PyQt6 (or pySide6) would make the installation process smoother, but I do not think that should be a requirement in terms of acceptance.

@JohannesBuchner do you have any further comments on the installation process?

----
Error installing pyqt5
```

Collecting PyQt5==5.15.2
  Downloading PyQt5-5.15.2.tar.gz (3.3 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [23 lines of output]
      pyproject.toml: line 7: using '[tool.sip.metadata]' to specify the project metadata is deprecated and will be removed in SIP v7.0.0, use '[project]' instead
      Traceback (most recent call last):
        File "/home/trygvrad/go-xrx/venv310/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
        File "/home/trygvrad/go-xrx/venv310/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/home/trygvrad/go-xrx/venv310/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 152, in prepare_metadata_for_build_wheel
          whl_basename = backend.build_wheel(metadata_directory, config_settings)
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/sipbuild/api.py", line 28, in build_wheel
          project = AbstractProject.bootstrap('wheel',
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/sipbuild/abstract_project.py", line 74, in bootstrap
          project.setup(pyproject, tool, tool_description)
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/sipbuild/project.py", line 608, in setup
          self.apply_user_defaults(tool)
        File "/tmp/pip-install-uy621jmq/pyqt5_8ee1a2e5dec54ffc91ada5ee74f8621c/project.py", line 63, in apply_user_defaults
          super().apply_user_defaults(tool)
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/pyqtbuild/project.py", line 51, in apply_user_defaults
          super().apply_user_defaults(tool)
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/sipbuild/project.py", line 237, in apply_user_defaults
          self.builder.apply_user_defaults(tool)
        File "/tmp/pip-build-env-wpw37r1d/overlay/lib/python3.10/site-packages/pyqtbuild/builder.py", line 50, in apply_user_defaults
          raise PyProjectOptionException('qmake',
      sipbuild.pyproject.PyProjectOptionException
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
---
Error opening in wsl running pyqt6
```
Failed to create wl_display (No such file or directory)
qt.qpa.plugin: Could not load the Qt platform plugin “wayland” in “” even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.
```
