### 2024-11-21 2c766d3
Updated paper and readme - reviewer suggestions
