+++ PAPER/paper.md
@@ -48,7 +48,8 @@ One notable example is the analysis of LaMnO<sub>3</sub>/SrTiO<sub>3</sub> thin-
 
 To the best of our knowledge, no existing software tool comprehensively addresses the general problem of RXR data analysis across a wide range of materials and conditions, e.g., thin films, oxides, and metals under x-ray energies from around 100~eV to 100~keV. Most of the currently available tools are designed for specific tasks and cannot be extensively applied to the diverse challenges encountered in RXR studies especially complex layer and interface characterization, overlapping resonances, and robust parameter estimation. For instance, tools like GenX [@bjorck2007_genx] and Motofit [@nelson2006_motofit], while effective for neutron and non-resonant x-ray reflectometry, lack the capability to model key aspects of RXR data, such as magnetic scattering contributions and energy-dependent changes in refractive indices, which are critical for interpreting element-specific interactions in complex material. Similarly, more modern tools like Refl1D [@refl1d_github] and BornAgain [@pospelov2020_bornagain] offer powerful features for modeling and fitting specific types of reflectometry data but still lack the flexibility and comprehensive approach needed for the broad range of challenges in RXR analysis.
 
-GO-RXR was developed to fill this gap by providing a comprehensive tool for RXR data analysis. It incorporates advanced global optimization algorithms that efficiently handle the high dimensionality and complexity of RXR datasets, avoiding common issues like local minima. GO-RXR’s ability to customize objective functions and integrate fitting techniques, such as total variation penalty methods, enables accurate modeling of complex electronic and magnetic structures in thin films. This makes it uniquely suited for a broad range of material systems and experimental conditions **[examples here]**.
+GO-RXR was developed to fill this gap by providing a comprehensive tool for RXR data analysis. It incorporates advanced global optimization algorithms that efficiently handle the high dimensionality and complexity of RXR datasets, avoiding common issues like local minima. GO-RXR's ability to customize objective functions and integrate fitting techniques, such as total variation penalty methods, enables accurate modeling of complex electronic and magnetic structures in thin films. This makes it uniquely suited for addressing the broad challenges in RXR data analysis across diverse material systems and experimental conditions.
+
 
 Additionally, GO-RXR features a user-friendly graphical interface with intuitive plotting options, enhancing data interpretation and visualization. This combination of advanced analytical capabilities and accessibility bridges the gap between experimental data collection and theoretical modeling, offering researchers a powerful tool for studying complex material systems in RXR.
 
