### 2024-11-21 #7165 @jpcurbelo
Hi @trygvrad. Thank you for your detailed feedback. We have made the following updates based on your suggestions:

1- Clarity in the Sentence on Suitability

- We removed the "[Examples here]" part and revised the sentence to:

  "This makes it uniquely suited for addressing the broad challenges in RXR data analysis across diverse material systems and experimental conditions."

- To avoid repetition, we did not add specific examples here as they are already provided in the previous paragraph and throughout the other sections.

2- README Updates:
- Created a new subsection, **Advanced Installation Notes and Known Issues**, to consolidate prior issues and include additional notes.
- Added issues/solutions for Ubuntu 24.04 and WSL on ARM64.
- Updated the instructions to use `python3.10 -m venv venv-go-rxr` for consistency across all setup instructions.

We hope these changes address your concerns and improve the clarity and usability of the documentation. Thank you again for your thorough review!
