### 2024-11-21 lucaskorol21/GO-RXR#43 @JohannesBuchner
Can you confirm that you tested, by clicking at each, and that works for you?
