### 2024-11-21 lucaskorol21/GO-RXR#43 @jpcurbelo
Hi, @JohannesBuchner. Yes, I can confirm that I have tested and the three options are working as expected. Please, take a look at the screencast and let us know whether you are facing any issues on your end. Tahnks.
[go-rxr_help_menu.webm](https://github.com/user-attachments/assets/2974f11f-cf9a-4d39-871d-6e1a904c50b2)
