### 2024-12-15 lucaskorol21/GO-RXR#43 @JohannesBuchner
Works for me with latest version
