### 2024-12-15 #7165 @JohannesBuchner
Today, installing on Ubuntu following the installation instructions with deadsnakes/ppa went smoothly. I looked over the paper once more, and have no remaining comments. I am happy to recommend publication.
