### 2025-01-02 #7165 @zhubonan
Hi @jpcurbelo, there are a few things to be done before we accept the paper. Please use the checklist above for a guide.
Do you and other authors have an ORCID? It would be useful to add those to the paper as well as when making the archive.

Please make a release of the software and record the version number, this number will be linked to the JOSS paper publication.
You can use Zenodo to archive the software code - this will generate a DOI for the code, and please let me know.
If you use Zenodo, please make sure the title and authors matches with that of the paper. It would be useful to also include an abstract. Note that if you link GitHub with Zenodo then the default title will be the repository's one which needs to be changed.
