### 2025-01-29 #7165 @jpcurbelo
Hi @zhubonan. Here are the updates:

- The author's ORCID numbers were updated in the paper.
- We made[the release v1.0.7](https://github.com/lucaskorol21/GO-RXR/tree/v1.0.7).
- We have created [the archive on Zenodo](https://zenodo.org/records/14758968) linking the release in the Github repo. We have also included an abstract.

Please, let us know if you'd need anything else from our end. Thanks.
