### 2025-02-04 #7165 @zhubonan
@jpcurbelo Thanks for the update. Just a few minor things:
1. there is an invalid DOI, I have made a PR: https://github.com/lucaskorol21/GO-RXR/pull/49
2. Please update the authors field in metadata of the Zenodo to match that of the paper. Please include the ORCIDs there as well.  You should be able to directly edit it without making a new release.
