### 2025-02-19 #7165 @jpcurbelodp
Hi @zhubonan. I just thought to let you know that all authors and their affiliations should be on [the Zenodo archive](https://zenodo.org/records/14758968) now. Please, let us know about the next steps and whether you'd need anything else from our end. Thanks.
