### 2025-03-02 #7165 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
