### 2025-03-13 #7165 @jpcurbelodp
Hi @Kevin-Mattheus-Moerman. Thanks for taking care of this and excuse the delay in answering to your request. Could, you, please, clarify a couple of points? Thanks.

We've noticed that you have updated the version to be v1.0.8 (I belive this was generated after the merge). It happens that release that was archived on Zenodo was [v1.0.7](https://github.com/lucaskorol21/GO-RXR/releases/tag/v1.0.7). @lucaskorol21, who is the repo author, had to make the release him self in order to match the repo on Github.

So, the question is whether should @lucaskorol21 create a new release with the corrected title, then, you will update the version here? We are not quite sure how to proceed and would like to avoid future complications. Thanks again.
