### 2025-03-13 #7165 @Kevin-Mattheus-Moerman
@jpcurbelodp you can use the archive for v1.0.7 or v1.0.8 (I updated as I assumed the latest would be best). You can go to your archive link, sign into ZENODO, then click edit and manually edit the archive title.

See this screen recording I just took:
https://github.com/user-attachments/assets/5d5d2aee-ce66-4167-b594-1bdc4ce44372

Let me know what version you go with.
